pyproject.toml
[tool.poetry]
name = "reminders-bot"
version = "0.1.0"
description = ""
authors = ["adrianvillanueva997 <adrianvillanueva997@gmail.com>"]
readme = "README.md"
packages = [{ include = "reminders_bot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = { extras = ["job-queue", "webhook"], version = "^20.4" }
rich = "^13.4.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.278"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
